+++ .gitignore
@@ -1,3 +1,4 @@
+client/node_modules/
 server/node_modules/
 server/lib/
 server/src/gql-types.ts
