+++ client/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "client",
+  "scripts": {
+    "build": "npm run clean && tsc && npm run lint",
+    "build:watch": "tsc --watch",
+    "lint": "tslint -c ../tslint.json  src/**/*.ts -t stylish",
+    "clean": "rm -rf lib"
+  },
+  "dependencies": {
+    "lit-html": "^0.7.1"
+  },
+  "devDependencies": {
+    "tslint": "^5.8.0",
+    "typescript": "^2.6.2"
+  }
+}
